treewide: refactor gallery window and use adwaita widgets instead of GtkBox
the initialization of the class was wrong to begin with. what it did was append a GtkBox on top of Adw.Application. while what should actually happen is to have an Application start an ApplicationWindow and then ToolbarView and Navigation etc
because of this, the app looked did not use any of the actual adwaita widgets at all and just appended stuff into a box. all classes still need heavy refactoring but this is a good starting point

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>